requirements.txt
langchain==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-google-genai
langchain-postgres==0.0.16
langchain-core==1.1.0
python-dotenv==1.1.0
langchain-experimental==0.4.1
unstructured[all-docs]==0.18.20
langchain-tavily==0.2.15
langgraph==1.0.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==1.0.2
langgraph-sdk
langsmith
psycopg2
PyYAML
# brew install libmagic poppler tesseract